sorting file locations
